fix edit and delete functionality

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -95,14 +95,27 @@ public class ImageController {
 
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
-    @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
-        Image image = imageService.getImage(imageId);
 
+    //bug - a non owner of image can also edit the image
+    // compare session user with the image user
+    @RequestMapping(value = "/editImage")
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
+        Image image = imageService.getImage(imageId);
+        User user=(User) session.getAttribute("loggeduser");
+        User owner=image.getUser();
+        if(user.getId() == owner.getId()){
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
         model.addAttribute("tags", tags);
-        return "images/edit";
+        return "images/edit";}
+        else{
+            String editError="Only the owner of the image can edit the image";
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("editError", editError);
+            return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -143,10 +156,25 @@ public class ImageController {
     //This controller method is called when the request pattern is of type 'deleteImage' and also the incoming request is of DELETE type
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
+
+    //bug - a non owner of image can also delete the image
+    // compare session user with the image user
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session,Model model) {
+        User user=(User) session.getAttribute("loggeduser");
+        Image image= imageService.getImage(imageId);
+        User owner=image.getUser();
+        if(user.getId() == owner.getId()){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else{
+            String deleteError="Only the owner of the image can delete the image";
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("deleteError", deleteError);
+            return "images/image";
+        }
     }
